# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Chamados — API REST
  version: 1.0.0
  description: >
    Contrato da API REST do Sistema de Chamados (help desk / service desk).
    Define autenticação (JWT), ciclo de vida dos chamados, comentários, anexos,
    histórico, gestão de domínios (status/categoria/prioridade/SLA), usuários e
    dashboard. Este documento é a fonte da verdade do contrato; a implementação
    (ASP.NET Core) e o schema físico (PostgreSQL via EF Core) seguem-no.
servers:
  - url: /api/v1
    description: Base path da API
...
security:
  - bearerAuth: []
paths:
...
  /usuarios/tecnicos:
    get:
      tags: [Usuários]
      summary: Lista técnicos ativos
      description: >
        Endpoint enxuto (sem paginação) usado para popular seletores de
        técnico, como o filtro da listagem de chamados. Distinto do
        CRUD administrativo em `/usuarios`.
      responses:
        '200':
          description: Lista de técnicos ativos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /usuarios/atribuiveis:
    get:
      tags: [Usuários]
      summary: Lista usuários que podem receber atribuição de chamado
      description: >
        Endpoint enxuto (sem paginação), usado para popular o seletor de
        `/chamados/{id}/atribuir`. Sempre inclui os técnicos ativos; se o
        usuário autenticado for Administrador, também inclui os
        administradores ativos. Distinto do CRUD administrativo em
        `/usuarios`.
      responses:
        '200':
          description: Lista de usuários atribuíveis
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /usuarios:
    get:
      tags: [Usuários]
      summary: Lista usuários (Administrador)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Página de usuários
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Usuários]
      summary: Cria um usuário (Administrador)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuarioCreate' }
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /usuarios/{id}/departamentos:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    put:
      tags: [Usuários]
      summary: Substitui os departamentos de um usuário (Administrador)
      description: >
        Substitui integralmente o vínculo N:N do usuário com departamentos
        (não é incremental — os `idsDepartamentos` enviados passam a ser a
        lista completa). Falha com 422 se algum id não corresponder a um
        departamento existente.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuarioDepartamentosRequest' }
      responses:
        '200':
          description: Usuário com departamentos atualizados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

  # ---------------------------------------------------------------- Departamentos (admin)
  /departamentos:
    get:
      tags: [Departamentos]
      summary: Lista departamentos
      parameters:
        - name: apenasAtivos
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Lista de departamentos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Departamento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Departamentos]
      summary: Cria um departamento (Administrador)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartamentoCreate' }
      responses:
        '201':
          description: Departamento criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Departamento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /departamentos/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    patch:
      tags: [Departamentos]
      summary: Atualiza nome/descrição de um departamento (Administrador)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartamentoUpdate' }
      responses:
        '200':
          description: Departamento atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Departamento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /departamentos/{id}/status:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    patch:
      tags: [Departamentos]
      summary: Ativa ou inativa um departamento (Administrador)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartamentoStatusRequest' }
      responses:
        '200':
          description: Departamento com status atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Departamento' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Domínios (admin)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      description: Número da página (base 1)
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Itens por página
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: Requisição malformada
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Não autenticado (token ausente/inválido)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Autenticado, mas sem permissão para o recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 'Conflito (ex.: e-mail já existe, transição de status inválida)'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Falha de validação dos campos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      description: Formato padronizado de erro para toda a API.
      required: [status, title]
      properties:
        status: { type: integer, example: 422 }
        title: { type: string, example: 'Falha de validação' }
        detail: { type: string, example: 'Um ou mais campos são inválidos.' }
        traceId: { type: string, example: '00-abc123...' }
        errors:
          type: object
          description: Mapa campo → lista de mensagens (para 422)
          additionalProperties:
            type: array
            items: { type: string }
      examples:
        - status: 422
          title: 'Falha de validação'
          errors:
            titulo: ['O título é obrigatório.']
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
        totalItems: { type: integer, example: 137 }
        totalPages: { type: integer, example: 7 }

    # ----- Auth -----
...
    Perfil:
      type: string
      description: Papel de acesso do usuário
      enum: [ADMINISTRADOR, TECNICO, CLIENTE]
    Usuario:
      type: object
      properties:
        id: { type: integer, format: int64 }
        nome: { type: string }
        email: { type: string, format: email }
        perfil: { $ref: '#/components/schemas/Perfil' }
        ativo: { type: boolean }
        criadoEm: { type: string, format: date-time }
        departamentos:
          type: array
          items: { $ref: '#/components/schemas/Departamento' }
          description: Departamentos aos quais o usuário pertence (relevante para perfil Técnico)
    UsuarioCreate:
      type: object
      required: [nome, email, senha, perfil]
      properties:
        nome: { type: string, maxLength: 120 }
        email: { type: string, format: email, maxLength: 160 }
        senha: { type: string, format: password, minLength: 6 }
        perfil: { $ref: '#/components/schemas/Perfil' }
...
    UsuarioPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Usuario' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    UsuarioDepartamentosRequest:
      type: object
      required: [idsDepartamentos]
      properties:
        idsDepartamentos:
          type: array
          items: { type: integer, format: int64 }

    # ----- Departamentos -----
    Departamento:
      type: object
      properties:
        id: { type: integer, format: int64 }
        nome: { type: string }
        descricao: { type: string }
        ativo: { type: boolean }
        criadoEm: { type: string, format: date-time }
    DepartamentoCreate:
      type: object
      required: [nome]
      properties:
        nome: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 255 }
    DepartamentoUpdate:
      type: object
      properties:
        nome: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 255 }
    DepartamentoStatusRequest:
      type: object
      required: [ativo]
      properties:
        ativo: { type: boolean }

    # ----- Domínios -----
...